backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.3.2
pymongo[srv]==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
certifi>=2023.0.0
